LayoutTests/fast/selectors/nth-child-of-never-matching-selector-expected.txt

Check the behavior of the ":nth-child(An+B of selectorList)" pseudo class when the selectorList does not or cannot match anything.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing ":nth-child(2 of :not(*))"
PASS document.querySelectorAll('#test-root :nth-child(2 of :not(*))').length is 0
PASS styledElements is []

Testing ":nth-child(n+2 of :not(*))"
PASS document.querySelectorAll('#test-root :nth-child(n+2 of :not(*))').length is 0
PASS styledElements is []

Testing ":nth-child(1 of :not(*))"
PASS document.querySelectorAll('#test-root :nth-child(1 of :not(*))').length is 0
PASS styledElements is []

Testing ":nth-child(n of :not(*))"
PASS document.querySelectorAll('#test-root :nth-child(n of :not(*))').length is 0
PASS styledElements is []

Testing ":nth-child(n-99999 of :not(*))"
PASS document.querySelectorAll('#test-root :nth-child(n-99999 of :not(*))').length is 0
PASS styledElements is []

Testing ":nth-child(n+2 of [attribute4])"
PASS document.querySelectorAll('#test-root :nth-child(n+2 of [attribute4])').length is 0
PASS styledElements is []

Testing ":nth-child(n+2 of [attribute1=value2])"
PASS document.querySelectorAll('#test-root :nth-child(n+2 of [attribute1=value2])').length is 0
PASS styledElements is []

Testing ":nth-child(n+2 of testcaseC)"
PASS document.querySelectorAll('#test-root :nth-child(n+2 of testcaseC)').length is 0
PASS styledElements is []

Testing ":nth-child(n+2 of .not-there)"
PASS document.querySelectorAll('#test-root :nth-child(n+2 of .not-there)').length is 0
PASS styledElements is []

Testing ":nth-child(n+2 of #not-there)"
PASS document.querySelectorAll('#test-root :nth-child(n+2 of #not-there)').length is 0
PASS styledElements is []

Testing ":nth-child(2 of :not(*), [attribute4], [attribute1=value2], testcaseC, .not-there, #not-there)"
PASS document.querySelectorAll('#test-root :nth-child(2 of :not(*), [attribute4], [attribute1=value2], testcaseC, .not-there, #not-there)').length is 0
PASS styledElements is []

PASS successfullyParsed is true

TEST COMPLETE
